+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Remove node classes from vsphere template used in `kubectl gs template cluster` command.
+
 ## [4.5.0] - 2024-11-15
 
 ### Changed